+++ packages/app/src/@generic/utils/hide-app-splash-screen.web.ts
@@ -1,0 +1,7 @@
+import { WebSplashElementId } from '../constants/web-splash.constant';
+
+export const hideAppSplashScreen = (): Promise<void> => {
+    document.getElementById(WebSplashElementId)?.remove();
+
+    return Promise.resolve();
+};