+++ packages/techniques/src/@generic/types/fish-base.type.ts
@@ -1,0 +1,8 @@
+import type { LineType } from './line.type';
+
+export type FishBaseType = {
+    readonly unitPositions: readonly number[];
+    readonly baseType: LineType;
+    readonly coverType: LineType;
+    readonly value: number;
+};