+++ packages/rating/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.11.4](https://github.com/vitalyiegorov/suuudokuuu/compare/v2.11.3...v2.11.4) (2026-09-02)
+
+### Performance Improvements
+
+- **app:** fast web first paint, smaller web and native bundles, desktop tools and hint panel polish ([#391](https://github.com/vitalyiegorov/suuudokuuu/issues/391)) ([6bf7797](https://github.com/vitalyiegorov/suuudokuuu/commit/6bf7797b88345903248ffcb1311f4d1e5468ea95))
+
+## [2.11.2](https://github.com/vitalyiegorov/suuudokuuu/compare/v2.11.1...v2.11.2) (2026-09-01)
+
+**Note:** Version bump only for package @suuudokuuu/rating
+
 ## [2.10.1](https://github.com/vitalyiegorov/suuudokuuu/compare/v2.10.0...v2.10.1) (2026-08-30)
 
 **Note:** Version bump only for package @suuudokuuu/rating